+++ src/Widgets/BasicWidgets.hpp
@@ -13,7 +13,7 @@ namespace OpenSim
 
 namespace osc
 {
+	void DrawComponentHoverTooltip(OpenSim::Component const&);
 	void DrawSelectOwnerMenu(osc::VirtualModelStatePair&, OpenSim::Component const&);
 	void DrawRequestOutputsMenu(osc::MainUIStateAPI&, OpenSim::Component const&);
-
 }