+++ CHANGELOG.md
@@ -39,6 +39,8 @@ All notable changes to this project will be documented here. The format is based
   - This is an advanced feature that has been added to the `tools` menu in the editor tab
   - This makes it easier to compare the performance of `SimTK::Integrator`s for a particular sim
 - Simulations now cancel faster, and hung simulations should be closeable in more situations (#244)
+- Right-clicking a component in the simulations tab now shows a context menu that enables users to
+  select which output(s) they want to plot (#240)
 
 
 ## [0.1.6] - 2022/05/09